Extract reviewed execution coordinator
