+++ control-plane/aegisops_control_plane/service.py
@@ -4,7 +4,6 @@ from contextlib import AbstractContextManager, contextmanager
 from collections import Counter
 from dataclasses import asdict, dataclass, fields, replace
 from datetime import datetime, timezone
-import hashlib
 import hmac
 import ipaddress
 import json
@@ -21,6 +20,7 @@ from .adapters.postgres import (
 from .adapters.shuffle import ShuffleActionAdapter
 from .adapters.wazuh import WazuhAlertAdapter
 from .config import RuntimeConfig
+from .execution_coordinator import ExecutionCoordinator
 from .models import (
     AITraceRecord,
     ActionExecutionRecord,
@@ -747,26 +747,6 @@ def _record_from_backup_payload(
         kwargs[field_info.name] = value
     return record_type(**kwargs)
 
-
-def _approved_payload_binding_hash(
-    *,
-    target_scope: Mapping[str, object],
-    approved_payload: Mapping[str, object],
-    execution_surface_type: str,
-    execution_surface_id: str,
-) -> str:
-    binding = _json_ready(
-        {
-            "approved_payload": approved_payload,
-            "execution_surface_id": execution_surface_id,
-            "execution_surface_type": execution_surface_type,
-            "target_scope": target_scope,
-        }
-    )
-    encoded = json.dumps(binding, separators=(",", ":"), sort_keys=True).encode("utf-8")
-    return hashlib.sha256(encoded).hexdigest()
-
-
 def _merge_reviewed_context(
     existing_context: Mapping[str, object] | None,
     incoming_context: Mapping[str, object] | None,
@@ -914,6 +894,7 @@ class AegisOpsControlPlaneService:
                 _recommendation_draft_snapshot_from_context
             ),
         )
+        self._execution_coordinator = ExecutionCoordinator(self)
 
     def describe_runtime(self) -> RuntimeSnapshot:
         return RuntimeSnapshot(
@@ -1294,86 +1275,13 @@ class AegisOpsControlPlaneService:
         delegation_issuer: str,
         evidence_ids: tuple[str, ...] = (),
     ) -> ActionExecutionRecord:
-        delegated_at = self._require_aware_datetime(delegated_at, "delegated_at")
-        action_request_id = self._require_non_empty_string(
-            action_request_id,
-            "action_request_id",
+        return self._execution_coordinator.delegate_approved_action_to_shuffle(
+            action_request_id=action_request_id,
+            approved_payload=approved_payload,
+            delegated_at=delegated_at,
+            delegation_issuer=delegation_issuer,
+            evidence_ids=evidence_ids,
         )
-        delegation_issuer = self._require_non_empty_string(
-            delegation_issuer,
-            "delegation_issuer",
-        )
-        normalized_payload = self._require_mapping(approved_payload, "approved_payload")
-        with self._store.transaction():
-            action_request, approval_decision = self._load_approved_delegation_context(
-                action_request_id=action_request_id,
-                approved_payload=normalized_payload,
-                delegated_at=delegated_at,
-                execution_surface_type="automation_substrate",
-                execution_surface_id="shuffle",
-                invalid_execution_surface_type_message=(
-                    "approved action request is not delegated through the automation "
-                    "substrate path"
-                ),
-                invalid_execution_surface_id_message=(
-                    "approved action request is not routed to the reviewed shuffle adapter"
-                ),
-                delegation_label="shuffle",
-            )
-            approval_decision_id = approval_decision.approval_decision_id
-            for existing in self._store.list(ActionExecutionRecord):
-                if (
-                    existing.action_request_id == action_request.action_request_id
-                    and existing.execution_surface_type == "automation_substrate"
-                    and existing.execution_surface_id == "shuffle"
-                    and existing.idempotency_key == action_request.idempotency_key
-                ):
-                    return existing
-
-            delegation_id = self._next_identifier("delegation")
-            receipt = self._shuffle.dispatch_approved_action(
-                delegation_id=delegation_id,
-                action_request_id=action_request.action_request_id,
-                approval_decision_id=approval_decision_id,
-                payload_hash=action_request.payload_hash,
-                idempotency_key=action_request.idempotency_key,
-                approved_payload=normalized_payload,
-                delegated_at=delegated_at,
-            )
-            provenance: dict[str, object] = {
-                "delegation_issuer": delegation_issuer,
-                "evidence_ids": evidence_ids,
-                "adapter": receipt.adapter,
-                "downstream_binding": {
-                    "approval_decision_id": receipt.approval_decision_id,
-                    "delegation_id": receipt.delegation_id,
-                    "payload_hash": receipt.payload_hash,
-                },
-            }
-            if receipt.base_url.strip() and receipt.base_url != "<set-me>":
-                provenance["adapter_base_url"] = receipt.base_url
-
-            execution = self.persist_record(
-                ActionExecutionRecord(
-                    action_execution_id=self._next_identifier("action-execution"),
-                    action_request_id=action_request.action_request_id,
-                    approval_decision_id=approval_decision_id,
-                    delegation_id=delegation_id,
-                    execution_surface_type=receipt.execution_surface_type,
-                    execution_surface_id=receipt.execution_surface_id,
-                    execution_run_id=receipt.execution_run_id,
-                    idempotency_key=action_request.idempotency_key,
-                    target_scope=action_request.target_scope,
-                    approved_payload=normalized_payload,
-                    payload_hash=action_request.payload_hash,
-                    delegated_at=delegated_at,
-                    expires_at=action_request.expires_at,
-                    provenance=provenance,
-                    lifecycle_state="queued",
-                )
-            )
-        self._emit_action_execution_delegated_event(execution)
-        return execution
 
     def delegate_approved_action_to_isolated_executor(
         self,
@@ -1384,81 +1292,13 @@ class AegisOpsControlPlaneService:
         delegation_issuer: str,
         evidence_ids: tuple[str, ...] = (),
     ) -> ActionExecutionRecord:
-        delegated_at = self._require_aware_datetime(delegated_at, "delegated_at")
-        action_request_id = self._require_non_empty_string(
-            action_request_id,
-            "action_request_id",
+        return self._execution_coordinator.delegate_approved_action_to_isolated_executor(
+            action_request_id=action_request_id,
+            approved_payload=approved_payload,
+            delegated_at=delegated_at,
+            delegation_issuer=delegation_issuer,
+            evidence_ids=evidence_ids,
         )
-        delegation_issuer = self._require_non_empty_string(
-            delegation_issuer,
-            "delegation_issuer",
-        )
-        normalized_payload = self._require_mapping(approved_payload, "approved_payload")
-        with self._store.transaction():
-            action_request, approval_decision = self._load_approved_delegation_context(
-                action_request_id=action_request_id,
-                approved_payload=normalized_payload,
-                delegated_at=delegated_at,
-                execution_surface_type="executor",
-                execution_surface_id="isolated-executor",
-                invalid_execution_surface_type_message=(
-                    "approved action request is not delegated through the isolated "
-                    "executor path"
-                ),
-                invalid_execution_surface_id_message=(
-                    "approved action request is not routed to the reviewed isolated executor"
-                ),
-                delegation_label="isolated executor",
-            )
-            approval_decision_id = approval_decision.approval_decision_id
-            for existing in self._store.list(ActionExecutionRecord):
-                if (
-                    existing.action_request_id == action_request.action_request_id
-                    and existing.execution_surface_type == "executor"
-                    and existing.execution_surface_id == "isolated-executor"
-                    and existing.idempotency_key == action_request.idempotency_key
-                ):
-                    return existing
-
-            delegation_id = self._next_identifier("delegation")
-            receipt = self._isolated_executor.dispatch_approved_action(
-                delegation_id=delegation_id,
-                action_request_id=action_request.action_request_id,
-                approval_decision_id=approval_decision_id,
-                payload_hash=action_request.payload_hash,
-                idempotency_key=action_request.idempotency_key,
-                approved_payload=normalized_payload,
-                delegated_at=delegated_at,
-            )
-            provenance: dict[str, object] = {
-                "delegation_issuer": delegation_issuer,
-                "evidence_ids": evidence_ids,
-                "adapter": receipt.adapter,
-            }
-            if receipt.base_url.strip() and receipt.base_url != "<set-me>":
-                provenance["adapter_base_url"] = receipt.base_url
-
-            execution = self.persist_record(
-                ActionExecutionRecord(
-                    action_execution_id=self._next_identifier("action-execution"),
-                    action_request_id=action_request.action_request_id,
-                    approval_decision_id=approval_decision_id,
-                    delegation_id=delegation_id,
-                    execution_surface_type=receipt.execution_surface_type,
-                    execution_surface_id=receipt.execution_surface_id,
-                    execution_run_id=receipt.execution_run_id,
-                    idempotency_key=action_request.idempotency_key,
-                    target_scope=action_request.target_scope,
-                    approved_payload=normalized_payload,
-                    payload_hash=action_request.payload_hash,
-                    delegated_at=delegated_at,
-                    expires_at=action_request.expires_at,
-                    provenance=provenance,
-                    lifecycle_state="queued",
-                )
-            )
-        self._emit_action_execution_delegated_event(execution)
-        return execution
 
     def evaluate_action_policy(self, action_request_id: str) -> ActionRequestRecord:
         action_request_id = self._require_non_empty_string(
@@ -2473,158 +2313,16 @@ class AegisOpsControlPlaneService:
         expires_at: datetime,
         action_request_id: str | None = None,
     ) -> ActionRequestRecord:
-        requester_identity = self._require_non_empty_string(
-            requester_identity,
-            "requester_identity",
+        return self._execution_coordinator.create_reviewed_action_request_from_advisory(
+            record_family=record_family,
+            record_id=record_id,
+            requester_identity=requester_identity,
+            recipient_identity=recipient_identity,
+            message_intent=message_intent,
+            escalation_reason=escalation_reason,
+            expires_at=expires_at,
+            action_request_id=action_request_id,
         )
-        recipient_identity = self._require_non_empty_string(
-            recipient_identity,
-            "recipient_identity",
-        )
-        message_intent = self._require_non_empty_string(
-            message_intent,
-            "message_intent",
-        )
-        escalation_reason = self._require_non_empty_string(
-            escalation_reason,
-            "escalation_reason",
-        )
-        expires_at = self._require_aware_datetime(expires_at, "expires_at")
-
-        with self._store.transaction():
-            context_snapshot = self.inspect_assistant_context(record_family, record_id)
-            self._require_phase19_case_scoped_advisory_read(context_snapshot)
-            recommendation_draft = self.render_recommendation_draft(
-                record_family,
-                record_id,
-            )
-            if recommendation_draft.recommendation_draft.get("status") != "ready":
-                raise ValueError(
-                    "reviewed advisory context is not ready for approval-bound action requests"
-                )
-
-            recommendation_id = self._require_single_recommendation_binding(
-                record_family=record_family,
-                record_id=record_id,
-                linked_recommendation_ids=recommendation_draft.linked_recommendation_ids,
-            )
-            case_id = self._require_single_linked_case_id(
-                recommendation_draft.linked_case_ids
-            )
-            case = self._require_phase19_operator_case(case_id)
-            if expires_at <= datetime.now(timezone.utc):
-                raise ValueError("expires_at must be in the future")
-
-            requested_payload = {
-                "action_type": "notify_identity_owner",
-                "recipient_identity": recipient_identity,
-                "message_intent": message_intent,
-                "escalation_reason": escalation_reason,
-                "source_record_family": record_family,
-                "source_record_id": record_id,
-                "recommendation_id": recommendation_id,
-                "case_id": case.case_id,
-                "alert_id": case.alert_id,
-                "finding_id": case.finding_id,
-                "linked_evidence_ids": recommendation_draft.linked_evidence_ids,
-            }
-            target_scope = {
-                "record_family": record_family,
-                "record_id": record_id,
-                "case_id": case.case_id,
-                "alert_id": case.alert_id,
-                "finding_id": case.finding_id,
-                "recipient_identity": recipient_identity,
-            }
-            payload_hash = _approved_payload_binding_hash(
-                target_scope=target_scope,
-                approved_payload=requested_payload,
-                execution_surface_type="automation_substrate",
-                execution_surface_id="shuffle",
-            )
-            requested_at = datetime.now(timezone.utc)
-            if expires_at <= requested_at:
-                raise ValueError("expires_at must be after requested_at")
-
-            idempotency_material = json.dumps(
-                {
-                    "payload_hash": payload_hash,
-                    "record_family": record_family,
-                    "record_id": record_id,
-                    "requester_identity": requester_identity,
-                    "expires_at": expires_at.isoformat(),
-                },
-                sort_keys=True,
-                separators=(",", ":"),
-            )
-            idempotency_key = (
-                "notify-identity-owner:"
-                + hashlib.sha256(idempotency_material.encode("utf-8")).hexdigest()
-            )
-            for existing in self._store.list(ActionRequestRecord):
-                if existing.idempotency_key == idempotency_key:
-                    self._emit_structured_event(
-                        logging.INFO,
-                        "action_request_reused",
-                        action_request_id=existing.action_request_id,
-                        action_type=existing.requested_payload.get("action_type"),
-                        lifecycle_state=existing.lifecycle_state,
-                        case_id=existing.case_id,
-                    )
-                    return existing
-
-            normalized_action_request_id = self._resolve_new_record_identifier(
-                ActionRequestRecord,
-                action_request_id,
-                "action_request_id",
-                "action-request",
-            )
-            created_request = self.persist_record(
-                ActionRequestRecord(
-                    action_request_id=normalized_action_request_id,
-                    approval_decision_id=None,
-                    case_id=case.case_id,
-                    alert_id=case.alert_id,
-                    finding_id=case.finding_id,
-                    idempotency_key=idempotency_key,
-                    target_scope=target_scope,
-                    payload_hash=payload_hash,
-                    requested_at=requested_at,
-                    expires_at=expires_at,
-                    lifecycle_state="pending_approval",
-                    requester_identity=requester_identity,
-                    requested_payload=requested_payload,
-                    policy_basis={
-                        "severity": "low",
-                        "target_scope": "single_identity",
-                        "action_reversibility": "reversible",
-                        "asset_criticality": "standard",
-                        "identity_criticality": "standard",
-                        "blast_radius": "single_target",
-                        "execution_constraint": "routine_allowed",
-                    },
-                    policy_evaluation={
-                        "approval_requirement": "human_required",
-                        "routing_target": "approval",
-                        "execution_surface_type": "automation_substrate",
-                        "execution_surface_id": "shuffle",
-                    },
-                )
-            )
-        self._emit_structured_event(
-            logging.INFO,
-            "action_request_created",
-            action_request_id=created_request.action_request_id,
-            action_type=created_request.requested_payload.get("action_type"),
-            requester_identity=created_request.requester_identity,
-            case_id=created_request.case_id,
-            alert_id=created_request.alert_id,
-            lifecycle_state=created_request.lifecycle_state,
-            expires_at=created_request.expires_at.isoformat()
-            if created_request.expires_at is not None
-            else None,
-        )
-        return created_request
 
     def attach_assistant_advisory_draft(
         self,
@@ -3857,246 +3555,14 @@ class AegisOpsControlPlaneService:
         compared_at: datetime,
         stale_after: datetime,
     ) -> ReconciliationRecord:
-        compared_at = self._require_aware_datetime(compared_at, "compared_at")
-        stale_after = self._require_aware_datetime(stale_after, "stale_after")
-        execution_surface_type = self._require_non_empty_string(
-            execution_surface_type,
-            "execution_surface_type",
-        )
-        execution_surface_id = self._require_non_empty_string(
-            execution_surface_id,
-            "execution_surface_id",
-        )
-        action_request = self._store.get(ActionRequestRecord, action_request_id)
-        if action_request is None:
-            raise LookupError(f"Missing action request {action_request_id!r}")
-        if action_request.lifecycle_state != "approved":
-            raise ValueError(
-                f"Action request {action_request_id!r} is not approved "
-                f"(state={action_request.lifecycle_state!r})"
-            )
-
-        require_binding_identifiers = (
-            execution_surface_type == "automation_substrate"
-            and execution_surface_id == "shuffle"
-        )
-        normalized_executions = self._normalize_observed_executions(
-            observed_executions,
-            require_binding_identifiers=require_binding_identifiers,
-        )
-        linked_execution_run_ids = tuple(
-            execution["execution_run_id"] for execution in normalized_executions
-        )
-        unique_execution_run_ids = tuple(dict.fromkeys(linked_execution_run_ids))
-        latest_execution = normalized_executions[-1] if normalized_executions else None
-        authoritative_execution = self._find_authoritative_action_execution(
-            action_request_id=action_request.action_request_id,
+        return self._execution_coordinator.reconcile_action_execution(
+            action_request_id=action_request_id,
             execution_surface_type=execution_surface_type,
             execution_surface_id=execution_surface_id,
-            execution_run_id=(
-                None if latest_execution is None else latest_execution["execution_run_id"]
-            ),
-            idempotency_key=(
-                action_request.idempotency_key
-                if latest_execution is None
-                else latest_execution["idempotency_key"]
-            ),
+            observed_executions=observed_executions,
+            compared_at=compared_at,
+            stale_after=stale_after,
         )
-
-        subject_linkage: dict[str, object] = {
-            "action_request_ids": (action_request.action_request_id,),
-            "execution_surface_types": (execution_surface_type,),
-            "execution_surface_ids": (execution_surface_id,),
-        }
-        if action_request.approval_decision_id is not None:
-            subject_linkage["approval_decision_ids"] = (
-                action_request.approval_decision_id,
-            )
-        if action_request.alert_id is not None:
-            subject_linkage["alert_ids"] = (action_request.alert_id,)
-        if action_request.case_id is not None:
-            subject_linkage["case_ids"] = (action_request.case_id,)
-        if action_request.finding_id is not None:
-            subject_linkage["finding_ids"] = (action_request.finding_id,)
-        if authoritative_execution is not None:
-            subject_linkage["action_execution_ids"] = (
-                authoritative_execution.action_execution_id,
-            )
-            subject_linkage["delegation_ids"] = (
-                authoritative_execution.delegation_id,
-            )
-            evidence_ids = self._merge_linked_ids(
-                authoritative_execution.provenance.get("evidence_ids"),
-                None,
-            )
-            if evidence_ids:
-                subject_linkage["evidence_ids"] = evidence_ids
-
-        ingest_disposition: str
-        lifecycle_state: str
-        mismatch_summary: str
-        execution_run_id: str | None = None
-        last_seen_at = action_request.requested_at
-
-        if latest_execution is None:
-            ingest_disposition = "missing"
-            lifecycle_state = "pending"
-            mismatch_summary = (
-                "missing downstream execution for approved action request correlation"
-            )
-        else:
-            execution_run_id = latest_execution["execution_run_id"]
-            last_seen_at = latest_execution["observed_at"]
-            observed_execution_surface_id = latest_execution["execution_surface_id"]
-            observed_idempotency_key = latest_execution["idempotency_key"]
-            observed_approval_decision_id = latest_execution.get("approval_decision_id")
-            observed_delegation_id = latest_execution.get("delegation_id")
-            observed_payload_hash = latest_execution.get("payload_hash")
-            expected_execution_run_id = (
-                None
-                if authoritative_execution is None
-                else authoritative_execution.execution_run_id
-            )
-            if last_seen_at < stale_after and compared_at >= stale_after:
-                ingest_disposition = "stale"
-                lifecycle_state = "stale"
-                mismatch_summary = "stale downstream execution observation requires refresh"
-            elif len(unique_execution_run_ids) > 1:
-                ingest_disposition = "duplicate"
-                lifecycle_state = "mismatched"
-                mismatch_summary = (
-                    "duplicate downstream executions observed for one approved request"
-                )
-            elif (
-                observed_execution_surface_id != execution_surface_id
-                or observed_idempotency_key != action_request.idempotency_key
-            ):
-                ingest_disposition = "mismatch"
-                lifecycle_state = "mismatched"
-                mismatch_summary = (
-                    "execution surface/idempotency mismatch between approved request and observed execution"
-                )
-            elif (
-                expected_execution_run_id is not None
-                and execution_run_id != expected_execution_run_id
-            ):
-                ingest_disposition = "mismatch"
-                lifecycle_state = "mismatched"
-                mismatch_summary = (
-                    "execution run identity mismatch between authoritative action execution "
-                    "and observed downstream execution"
-                )
-            elif (
-                authoritative_execution is not None
-                and authoritative_execution.execution_surface_type == "automation_substrate"
-                and authoritative_execution.execution_surface_id == "shuffle"
-                and (
-                    observed_approval_decision_id
-                    != authoritative_execution.approval_decision_id
-                    or observed_delegation_id != authoritative_execution.delegation_id
-                    or observed_payload_hash != authoritative_execution.payload_hash
-                )
-            ):
-                ingest_disposition = "mismatch"
-                lifecycle_state = "mismatched"
-                mismatch_summary = (
-                    "approved binding mismatch between authoritative action execution "
-                    "and observed downstream execution"
-                )
-            else:
-                ingest_disposition = "matched"
-                lifecycle_state = "matched"
-                mismatch_summary = (
-                    "matched approved action request to reviewed execution run"
-                )
-
-        if (
-            authoritative_execution is not None
-            and latest_execution is not None
-            and ingest_disposition == "matched"
-        ):
-            reconciled_lifecycle_state = self._action_execution_lifecycle_from_status(
-                latest_execution.get("status"),
-                authoritative_execution.lifecycle_state,
-            )
-            if reconciled_lifecycle_state != authoritative_execution.lifecycle_state:
-                authoritative_execution = self.persist_record(
-                    ActionExecutionRecord(
-                        action_execution_id=authoritative_execution.action_execution_id,
-                        action_request_id=authoritative_execution.action_request_id,
-                        approval_decision_id=authoritative_execution.approval_decision_id,
-                        delegation_id=authoritative_execution.delegation_id,
-                        execution_surface_type=authoritative_execution.execution_surface_type,
-                        execution_surface_id=authoritative_execution.execution_surface_id,
-                        execution_run_id=authoritative_execution.execution_run_id,
-                        idempotency_key=authoritative_execution.idempotency_key,
-                        target_scope=authoritative_execution.target_scope,
-                        approved_payload=authoritative_execution.approved_payload,
-                        payload_hash=authoritative_execution.payload_hash,
-                        delegated_at=authoritative_execution.delegated_at,
-                        expires_at=authoritative_execution.expires_at,
-                        provenance=authoritative_execution.provenance,
-                        lifecycle_state=reconciled_lifecycle_state,
-                    )
-                )
-
-        reconciliation = self.persist_record(
-            ReconciliationRecord(
-                reconciliation_id=self._next_identifier("reconciliation"),
-                subject_linkage=subject_linkage,
-                alert_id=action_request.alert_id,
-                finding_id=action_request.finding_id,
-                analytic_signal_id=None,
-                execution_run_id=execution_run_id,
-                linked_execution_run_ids=linked_execution_run_ids,
-                correlation_key=self._build_action_execution_reconciliation_key(
-                    action_request=action_request,
-                    execution_surface_type=execution_surface_type,
-                    execution_surface_id=execution_surface_id,
-                    authoritative_execution=authoritative_execution,
-                ),
-                first_seen_at=action_request.requested_at,
-                last_seen_at=last_seen_at,
-                ingest_disposition=ingest_disposition,
-                mismatch_summary=mismatch_summary,
-                compared_at=compared_at,
-                lifecycle_state=lifecycle_state,
-            )
-        )
-        self._emit_structured_event(
-            logging.INFO,
-            "action_execution_reconciled",
-            reconciliation_id=reconciliation.reconciliation_id,
-            action_request_id=action_request.action_request_id,
-            execution_surface_type=execution_surface_type,
-            execution_surface_id=execution_surface_id,
-            ingest_disposition=reconciliation.ingest_disposition,
-            lifecycle_state=reconciliation.lifecycle_state,
-            execution_run_id=reconciliation.execution_run_id,
-        )
-        return reconciliation
-
-    def _build_action_execution_reconciliation_key(
-        self,
-        *,
-        action_request: ActionRequestRecord,
-        execution_surface_type: str,
-        execution_surface_id: str,
-        authoritative_execution: ActionExecutionRecord | None,
-    ) -> str:
-        components = [action_request.action_request_id]
-        if action_request.approval_decision_id is not None:
-            components.append(action_request.approval_decision_id)
-        if authoritative_execution is not None:
-            components.append(authoritative_execution.delegation_id)
-        components.extend(
-            (
-                execution_surface_type,
-                execution_surface_id,
-                action_request.idempotency_key,
-            )
-        )
-        return ":".join(components)
 
     def _normalize_action_policy_basis(
         self,
@@ -5066,218 +4532,9 @@ class AegisOpsControlPlaneService:
             return substrate_detection_record_id
         return f"{namespaced_prefix}{substrate_detection_record_id}"
 
-    def _normalize_observed_executions(
-        self,
-        observed_executions: tuple[Mapping[str, object], ...],
-        *,
-        require_binding_identifiers: bool = False,
-    ) -> tuple[dict[str, object], ...]:
-        normalized: list[dict[str, object]] = []
-        for execution in observed_executions:
-            execution_run_id = execution.get("execution_run_id")
-            execution_surface_id = execution.get("execution_surface_id")
-            idempotency_key = execution.get("idempotency_key")
-            observed_at = execution.get("observed_at")
-            approval_decision_id = execution.get("approval_decision_id")
-            delegation_id = execution.get("delegation_id")
-            payload_hash = execution.get("payload_hash")
-            if not isinstance(execution_run_id, str):
-                raise ValueError("observed execution must include string execution_run_id")
-            if not isinstance(execution_surface_id, str):
-                raise ValueError(
-                    "observed execution must include string execution_surface_id"
-                )
-            if not isinstance(idempotency_key, str):
-                raise ValueError("observed execution must include string idempotency_key")
-            if not isinstance(observed_at, datetime):
-                raise ValueError("observed execution must include datetime observed_at")
-            observed_at = AegisOpsControlPlaneService._require_aware_datetime(
-                observed_at,
-                "observed_at",
-            )
-            if require_binding_identifiers:
-                if not isinstance(approval_decision_id, str):
-                    raise ValueError(
-                        "observed execution must include string approval_decision_id"
-                    )
-                if not isinstance(delegation_id, str):
-                    raise ValueError("observed execution must include string delegation_id")
-                if not isinstance(payload_hash, str):
-                    raise ValueError("observed execution must include string payload_hash")
-            normalized.append(
-                {
-                    "execution_run_id": execution_run_id,
-                    "execution_surface_id": execution_surface_id,
-                    "idempotency_key": idempotency_key,
-                    "observed_at": observed_at,
-                    "approval_decision_id": approval_decision_id,
-                    "delegation_id": delegation_id,
-                    "payload_hash": payload_hash,
-                    "status": execution.get("status"),
-                }
-            )
-
-        normalized.sort(key=lambda execution: execution["observed_at"])
-        return tuple(normalized)
-
-    def _find_authoritative_action_execution(
-        self,
-        *,
-        action_request_id: str,
-        execution_surface_type: str,
-        execution_surface_id: str,
-        execution_run_id: str | None,
-        idempotency_key: str,
-    ) -> ActionExecutionRecord | None:
-        matches = [
-            execution
-            for execution in self._store.list(ActionExecutionRecord)
-            if (
-                execution.action_request_id == action_request_id
-                and execution.execution_surface_type == execution_surface_type
-                and execution.execution_surface_id == execution_surface_id
-                and execution.idempotency_key == idempotency_key
-            )
-        ]
-        if execution_run_id is not None:
-            for execution in matches:
-                if execution.execution_run_id == execution_run_id:
-                    return execution
-        return matches[0] if matches else None
-
-    @staticmethod
-    def _action_execution_lifecycle_from_status(
-        status: object,
-        current_lifecycle_state: str,
-    ) -> str:
-        if not isinstance(status, str):
-            return current_lifecycle_state
-
-        normalized_status = status.strip().lower()
-        if normalized_status in {"queued", "pending"}:
-            return "queued"
-        if normalized_status in {"running", "in_progress"}:
-            return "running"
-        if normalized_status in {"success", "succeeded", "completed"}:
-            return "succeeded"
-        if normalized_status in {"failed", "error"}:
-            return "failed"
-        if normalized_status in {"canceled", "cancelled"}:
-            return "canceled"
-        return current_lifecycle_state
-
     @staticmethod
     def _next_identifier(prefix: str) -> str:
         return f"{prefix}-{uuid.uuid4()}"
-
-    @staticmethod
-    def _require_exact_approved_payload_binding(
-        *,
-        action_request: ActionRequestRecord,
-        approval_decision: ApprovalDecisionRecord,
-        approved_payload: Mapping[str, object],
-        execution_surface_type: str,
-        execution_surface_id: str,
-    ) -> None:
-        if approval_decision.action_request_id != action_request.action_request_id:
-            raise ValueError(
-                "approved payload binding does not match approved action request and approval decision"
-            )
-        if approval_decision.target_snapshot != action_request.target_scope:
-            raise ValueError(
-                "approved payload binding does not match approved action request and approval decision"
-            )
-
-        payload_hash = _approved_payload_binding_hash(
-            target_scope=action_request.target_scope,
-            approved_payload=approved_payload,
-            execution_surface_type=execution_surface_type,
-            execution_surface_id=execution_surface_id,
-        )
-        if (
-            payload_hash != action_request.payload_hash
-            or payload_hash != approval_decision.payload_hash
-        ):
-            raise ValueError(
-                "approved payload binding does not match approved action request and approval decision"
-            )
-
-    @staticmethod
-    def _require_exact_approved_expiry_binding(
-        *,
-        action_request: ActionRequestRecord,
-        approval_decision: ApprovalDecisionRecord,
-        delegated_at: datetime,
-        delegation_label: str,
-    ) -> None:
-        if approval_decision.approved_expires_at != action_request.expires_at:
-            raise ValueError("approved expiry window does not match action request expiry")
-        if (
-            approval_decision.approved_expires_at is not None
-            and delegated_at > approval_decision.approved_expires_at
-        ):
-            raise ValueError(
-                f"Action request {action_request.action_request_id!r} expired before {delegation_label} delegation"
-            )
-
-    def _load_approved_delegation_context(
-        self,
-        *,
-        action_request_id: str,
-        approved_payload: Mapping[str, object],
-        delegated_at: datetime,
-        execution_surface_type: str,
-        execution_surface_id: str,
-        invalid_execution_surface_type_message: str,
-        invalid_execution_surface_id_message: str,
-        delegation_label: str,
-    ) -> tuple[ActionRequestRecord, ApprovalDecisionRecord]:
-        action_request = self._store.get(ActionRequestRecord, action_request_id)
-        if action_request is None:
-            raise LookupError(f"Missing action request {action_request_id!r}")
-        if action_request.lifecycle_state != "approved":
-            raise ValueError(
-                f"Action request {action_request_id!r} is not approved "
-                f"(state={action_request.lifecycle_state!r})"
-            )
-        approval_decision_id = self._require_non_empty_string(
-            action_request.approval_decision_id,
-            "action_request.approval_decision_id",
-        )
-        approval_decision = self._store.get(ApprovalDecisionRecord, approval_decision_id)
-        if approval_decision is None:
-            raise LookupError(
-                f"Missing approval decision {approval_decision_id!r} for action request "
-                f"{action_request_id!r}"
-            )
-        if approval_decision.lifecycle_state != "approved":
-            raise ValueError(
-                f"Approval decision {approval_decision_id!r} is not approved "
-                f"(state={approval_decision.lifecycle_state!r})"
-            )
-        if approval_decision.payload_hash != action_request.payload_hash:
-            raise ValueError(
-                "approval decision payload_hash does not match action request payload_hash"
-            )
-        policy_evaluation = action_request.policy_evaluation
-        if policy_evaluation.get("execution_surface_type") != execution_surface_type:
-            raise ValueError(invalid_execution_surface_type_message)
-        if policy_evaluation.get("execution_surface_id") != execution_surface_id:
-            raise ValueError(invalid_execution_surface_id_message)
-        self._require_exact_approved_payload_binding(
-            action_request=action_request,
-            approval_decision=approval_decision,
-            approved_payload=approved_payload,
-            execution_surface_type=execution_surface_type,
-            execution_surface_id=execution_surface_id,
-        )
-        self._require_exact_approved_expiry_binding(
-            action_request=action_request,
-            approval_decision=approval_decision,
-            delegated_at=delegated_at,
-            delegation_label=delegation_label,
-        )
-        return action_request, approval_decision
 
     @staticmethod
     def _alert_review_state(alert: AlertRecord) -> str:
